pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-money-movement-poc"
version = "0.1.0"
description = "Bootstrap services for the agentic domestic money movement proof of concept."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0,<2.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "mcp>=1.0.0,<2.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.3,<4.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.4,<9.0.0",
]

[tool.setuptools]
packages = [
  "context_memory_service",
  "provenance_service",
  "orchestrator_api",
  "capability_gateway",
  "workflow_worker",
  "policy_service",
]

[tool.setuptools.package-dir]
context_memory_service = "services/context-memory-service/src/context_memory_service"
provenance_service = "services/provenance-service/src/provenance_service"
orchestrator_api = "apps/orchestrator-api/src/orchestrator_api"
capability_gateway = "apps/capability-gateway/src/capability_gateway"
workflow_worker = "services/workflow-worker/src/workflow_worker"
policy_service = "apps/policy-service/src/policy_service"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
  "services/context-memory-service/tests",
  "services/provenance-service/tests",
  "apps/orchestrator-api/tests",
  "apps/capability-gateway/tests",
  "services/workflow-worker/tests",
  "apps/policy-service/tests",
]
